+++ packages/transaction-controller/src/utils/provider.test.ts
@@ -1,4 +1,5 @@
 import type { NetworkClientId, Provider } from '@metamask/network-controller';
+import { RpcEndpointType } from '@metamask/network-controller';
 import type { Hex } from '@metamask/utils';
 
 import type { TransactionControllerMessenger } from '../TransactionController';
@@ -9,6 +10,7 @@ describe('provider utils', () => {
   const mockProvider = {
     request: requestMock,
   } as unknown as Provider;
+  const chainIdMock = '0xa4b1' as Hex;
 
   let messengerCallMock: jest.Mock;
   let messengerMock: TransactionControllerMessenger;
@@ -20,7 +22,23 @@ describe('provider utils', () => {
       .fn()
       .mockImplementation((action: string, ...args: unknown[]) => {
         if (action === 'NetworkController:getNetworkClientById') {
-          return { provider: mockProvider };
+          const networkClientId = args[0] as NetworkClientId;
+
+          return {
+            configuration: {
+              chainId: chainIdMock,
+              rpcEndpoints: [
+                {
+                  networkClientId,
+                  type:
+                    networkClientId === 'infuraNetworkClientId'
+                      ? RpcEndpointType.Infura
+                      : RpcEndpointType.Custom,
+                },
+              ],
+            },
+            provider: mockProvider,
+          };
         }
 
         if (action === 'NetworkController:findNetworkClientIdByChainId') {
@@ -106,6 +124,26 @@ describe('provider utils', () => {
           params: ['0x123', 'latest'],
         }),
       ).rejects.toBe(error);
+      expect(error.message).toBe(
+        'RPC 0xa4b1 Custom eth_getBalance: RPC failed',
+      );
+    });
+
+    it('identifies Infura provider.request errors', async () => {
+      const error = new Error('Unauthorized.');
+      requestMock.mockRejectedValue(error);
+
+      await expect(
+        rpcRequest({
+          messenger: messengerMock,
+          networkClientId: 'infuraNetworkClientId' as NetworkClientId,
+          method: 'eth_getBalance',
+          params: ['0x123', 'latest'],
+        }),
+      ).rejects.toBe(error);
+      expect(error.message).toBe(
+        'RPC 0xa4b1 Infura eth_getBalance: Unauthorized.',
+      );
     });
 
     it('works when params are undefined', async () => {