+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Add RPC method, chain ID, and endpoint type context to transaction pay provider errors ([#0000](https://github.com/MetaMask/core/pull/0000))
 - Bump `@metamask/keyring-controller` from `^27.0.0` to `^27.1.0` ([#9129](https://github.com/MetaMask/core/pull/9129))
 
 ### Fixed